remove page view tracking in favor of jetpack
