--- assets/js/extensions/jetpack/woocommerce-analytics/constants.ts
@@ -1,2 +1,0 @@
-export const namespace = 'jetpack-woocommerce-analytics';
-export const actionPrefix = 'experimental__woocommerce_blocks';
